requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
dj-rest-auth==7.0.1
Django==5.1.5
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.41
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.4
six==1.17.0
smmap==5.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
